+++ cfdm/core/abstract/properties.py
@@ -38,11 +38,16 @@ class Properties(with_metaclass(abc.ABCMeta, Container)):
         initialization By default parameters are deep copied.
 
         '''
-        super().__init__()
+        super().__init__(source=source, copy=copy)
         
         if source is not None:
-            properties = source.properties()
-        elif properties is None:
+            try:
+                properties = source.properties()
+            except AttributeError:
+                properties = None
+        #--- End: if
+
+        if properties is None:
             properties = {}
         
         self._set_component('properties', properties, copy=False)
@@ -77,24 +82,24 @@ class Properties(with_metaclass(abc.ABCMeta, Container)):
         return out.copy()
     #--- End: def
 
-    def copy(self):
-        '''Return a deep copy.
-
-``f.copy()`` is equivalent to ``copy.deepcopy(f)``.
-
-.. versionadded:: 1.7.0
-
-:Returns:
-
-        The deep copy.
-
-**Examples:**
-
->>> g = f.copy()
-
-        '''
-        return type(self)(source=self, copy=True)
-    #--- End: def
+#    def copy(self):
+#        '''Return a deep copy.
+#
+#``f.copy()`` is equivalent to ``copy.deepcopy(f)``.
+#
+#.. versionadded:: 1.7.0
+#
+#:Returns:
+#
+#        The deep copy.
+#
+#**Examples:**
+#
+#>>> g = f.copy()
+#
+#        '''
+#        return type(self)(source=self, copy=True)
+#    #--- End: def
 
     def del_property(self, prop, default=ValueError()):
         '''Remove a property.